performance improvements and simplifications
