+++ bench.ts
@@ -1,55 +1,31 @@
 import * as denoRedis from "@db/redis";
 import { Redis } from "ioredis";
 import { createClient } from "redis";
-import { RedisClient as DenoRedisClient } from "./deno.ts";
-import { RedisClient as WebRedisClient } from "./web.ts";
+import { RedisClient } from "./mod.ts";
 
 const hostname = "127.0.0.1";
 const port = 6379;
 
 const redisConn = await Deno.connect({ hostname, port });
-const denoRedisClient = new DenoRedisClient(redisConn);
-const webRedisClient = new WebRedisClient(redisConn);
+const redisClient = new RedisClient(redisConn);
 const denoRedisConn = await denoRedis.connect({ hostname, port });
 const ioRedis = new Redis();
 
 const nodeRedisClient = await createClient().connect();
 
 Deno.bench({
-  name: "@iuioiua/redis (deno)",
-  ignore: true,
+  name: "@iuioiua/redis",
   baseline: true,
   async fn() {
-    await denoRedisClient.sendCommand(["PING"]);
+    await redisClient.sendCommand(["PING"]);
 
-    await denoRedisClient.sendCommand(["SET", "mykey", "Hello"]);
-    await denoRedisClient.sendCommand(["GET", "mykey"]);
+    await redisClient.sendCommand(["SET", "mykey", "Hello"]);
+    await redisClient.sendCommand(["GET", "mykey"]);
 
-    await denoRedisClient.sendCommand(["HSET", "hash", "a", "foo", "b", "bar"]);
-    await denoRedisClient.sendCommand(["HGETALL", "hash"]);
+    await redisClient.sendCommand(["HSET", "hash", "a", "foo", "b", "bar"]);
+    await redisClient.sendCommand(["HGETALL", "hash"]);
 
-    await denoRedisClient.pipelineCommands([
-      ["INCR", "X"],
-      ["INCR", "X"],
-      ["INCR", "X"],
-      ["INCR", "X"],
-    ]);
-  },
-});
-
-Deno.bench({
-  name: "@iuioiua/redis (web)",
-  baseline: true,
-  async fn() {
-    await webRedisClient.sendCommand(["PING"]);
-
-    await webRedisClient.sendCommand(["SET", "mykey", "Hello"]);
-    await webRedisClient.sendCommand(["GET", "mykey"]);
-
-    await webRedisClient.sendCommand(["HSET", "hash", "a", "foo", "b", "bar"]);
-    await webRedisClient.sendCommand(["HGETALL", "hash"]);
-
-    await webRedisClient.pipelineCommands([
+    await redisClient.pipelineCommands([
       ["INCR", "X"],
       ["INCR", "X"],
       ["INCR", "X"],
