+++ deno.json
@@ -1,11 +1,7 @@
 {
   "name": "@iuioiua/redis",
   "version": "1.0.1",
-  "exports": {
-    "./": "./mod.ts",
-    "./deno": "./deno.ts",
-    "./web": "./web.ts"
-  },
+  "exports": "./mod.ts",
   "imports": {
     "@db/redis": "jsr:@db/redis@^0.38.0",
     "@std/assert": "jsr:@std/assert@^1.0.11",
@@ -19,7 +15,7 @@
   },
   "tasks": {
     "redis:start": "redis-server --save \"\" --appendonly no --daemonize yes",
-    "test": "deno test --allow-net --trace-leaks --coverage --doc --parallel --clean web_test.ts",
+    "test": "deno test --allow-net --trace-leaks --coverage --doc --parallel --clean",
     "test:dev": "deno task redis:start && deno task test ; redis-cli SHUTDOWN",
     "bench": "deno bench --allow-net --allow-env",
     "bench:dev": "deno task redis:start && deno task bench",
